requirements.txt
django
djangorestframework
django-rest-knox
markdown
djangorestframework-simplejwt